Extract test for ordering new bentos
